refactoring: Async cache, erp service, backend monitoring
The distributed cache interface is not asynchronous when accesing the
underlying redis connection.

The erp microservice now uses a background service to release the orders
that come int via call center and web shop.

The backend microservice now uses a monitoring background service that is
querying the distributed cache for the current values of the stocks, the
queue counts and delays.