ui: Fix stocks column height calculation and distribution
The stocs columns had two bugs. One was the missing `$` for accessing the
variable. After fixing that, the second bug emerged, where the distribution
of the different steps were still referencing the default `100px` height
value.

So each pecentage step needs its own media query to compensate for this
fact.