ui: Increase size on bigger screens and introduce query param
The first showcase revealed that the size of the nodes is still too small.
Let's add a debug mode to better test the code in realtime due to the
websocket issues when running locally.

With media queries it is just not possible to change properties in angular
components. So let's add a query parameter that can be used to change the
maximum queue size for smaller screens.